[BE011] Update Send Email for Verify Account Version 2
