+++ src/main/java/com/tttm/birdfarmshop/Controller/EmailController.java
@@ -3,11 +3,9 @@ package com.tttm.birdfarmshop.Controller;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tttm.birdfarmshop.Exception.CustomException;
-import com.tttm.birdfarmshop.Models.User;
-import com.tttm.birdfarmshop.Service.CodeStorageService;
 import com.tttm.birdfarmshop.Service.MailService;
 import com.tttm.birdfarmshop.DTO.MailDTO;
-import jakarta.servlet.http.HttpSession;
+import com.tttm.birdfarmshop.Utils.Response.AuthenticationResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,9 +19,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/email")
 public class EmailController {
     private final MailService mailService;
-    private final CodeStorageService codeStorageService;
+
     @PostMapping("/forgotPassword")
-    public ResponseEntity<String> ForgotPassword(@RequestBody String json) throws CustomException
+    public ResponseEntity<AuthenticationResponse> ForgotPassword(@RequestBody String json) throws CustomException
     {
         try
         {
@@ -33,50 +31,7 @@ public class EmailController {
         }
         catch (JsonProcessingException ex)
         {
-            return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_IMPLEMENTED);
-        }
-    }
-
-    @PostMapping("/sendCode")
-    public ResponseEntity<String> sendCode(@RequestBody String json, HttpSession session) throws CustomException
-    {
-        try
-        {
-            ObjectMapper mapper = new ObjectMapper();
-            MailDTO dto = mapper.readValue(json, MailDTO.class);
-
-            User user = (User) session.getAttribute(dto.getEmail());
-            String token = (String) session.getAttribute(user.toString());
-            if(token != null)
-            {
-                // Generate Code to Send Mail
-                String code = mailService.SendCode(dto.getEmail());
-                dto.setCode(code);
-
-                // Store Code in Session exist in 60 seconds
-                codeStorageService.storeCodeInSession(dto, session);
-                return new ResponseEntity<>(code, HttpStatus.OK);
-            }
-        }
-        catch (JsonProcessingException ex)
-        {
-            return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_IMPLEMENTED);
-        }
-        return new ResponseEntity<>("Empty Register", HttpStatus.OK);
-    }
-
-    @PostMapping("/verifyCode")
-    public ResponseEntity<String> verifyCode(@RequestBody String json, HttpSession session) throws CustomException
-    {
-        try
-        {
-            ObjectMapper mapper = new ObjectMapper();
-            MailDTO dto = mapper.readValue(json, MailDTO.class);
-            return new ResponseEntity<>(codeStorageService.getCodeFromSession(dto, session), HttpStatus.OK);
-        }
-        catch (JsonProcessingException ex)
-        {
-            return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_IMPLEMENTED);
+            return new ResponseEntity<>(new AuthenticationResponse(ex.getMessage()), HttpStatus.NOT_IMPLEMENTED);
         }
     }
