+++ src/main/java/com/tttm/birdfarmshop/Service/AuthenticationService.java
@@ -5,10 +5,9 @@ import com.tttm.birdfarmshop.Models.User;
 import com.tttm.birdfarmshop.Utils.Request.AuthenticationRequest;
 import com.tttm.birdfarmshop.Utils.Response.AuthenticationResponse;
 import com.tttm.birdfarmshop.Exception.CustomException;
+import com.tttm.birdfarmshop.Utils.Response.MessageResponse;
 import jakarta.servlet.http.HttpSession;
 
 public interface AuthenticationService {
-    AuthenticationResponse register(User dto, HttpSession session) throws CustomException;
-
    AuthenticationResponse login(AuthenticationRequest dto) throws CustomException;
 }
