+++ src/main/java/com/tttm/birdfarmshop/Service/CodeStorageService.java
@@ -1,10 +1,15 @@
 package com.tttm.birdfarmshop.Service;
 
 import com.tttm.birdfarmshop.DTO.MailDTO;
+import com.tttm.birdfarmshop.Exception.CustomException;
+import com.tttm.birdfarmshop.Models.User;
+import com.tttm.birdfarmshop.Utils.Response.AuthenticationResponse;
+import com.tttm.birdfarmshop.Utils.Response.MessageResponse;
 import jakarta.servlet.http.HttpSession;
 
 public interface CodeStorageService {
-     void storeCodeInSession(MailDTO dto, HttpSession session);
-     String getCodeFromSession(MailDTO dto, HttpSession session);
+     AuthenticationResponse getCodeFromSession(MailDTO dto, HttpSession session);
+
+     MessageResponse register(User dto, HttpSession session) throws CustomException;
 
 }
