+++ src/main/java/com/tttm/birdfarmshop/Utils/Response/MessageResponse.java
@@ -1,0 +1,14 @@
+package com.tttm.birdfarmshop.Utils.Response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class MessageResponse {
+    private String message;
+}
